requirements.txt
python-telegram-bot>=21.0
redis>=5.0
psycopg2-binary>=2.9
fastapi>=0.110
uvicorn>=0.30
python-dotenv>=1.0
httpx>=0.27
